@@ -6,7 +6,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.5.4</version>
+		<version>3.0.5</version>
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>digitalinnovation.one</groupId>
@@ -16,7 +16,7 @@
 	<description>Exemplos de Padrões de Projeto em Spring</description>
 	
 	<properties>
-		<java.version>11</java.version>
+		<java.version>17</java.version>
 		<openapi.version>1.5.10</openapi.version>
 		<spring-cloud.version>2020.0.3</spring-cloud.version>
 	</properties>
@@ -38,11 +38,11 @@
 			<artifactId>spring-cloud-starter-openfeign</artifactId>
 		</dependency>
 		<!-- OpenAPI/Swagger (adicionado manualmente) -->
-		<dependency>
+		<!--<dependency>
 			<groupId>org.springdoc</groupId>
 			<artifactId>springdoc-openapi-ui</artifactId>
 			<version>${openapi.version}</version>
-		</dependency>
+		</dependency>-->
 
 		<!-- H2 Database (adicionado via Spring Initializr) -->
 		<dependency>
@@ -55,6 +55,11 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.kafka</groupId>
+			<artifactId>spring-kafka</artifactId>
+			<version>3.0.5</version>
 		</dependency>
 	</dependencies>
 
@@ -78,5 +83,4 @@
 			</plugin>
 		</plugins>
 	</build>
-
 </project>
